Finish background service

+++ MostQuotation/Models/DollarQuotation.cs
@@ -2,7 +2,7 @@
 {
     public class DollarQuotation
     {
-        public DollarQuotation(DateTime dateTime, double quotation)
+        public DollarQuotation(DateTime dateTime, float quotation)
         {
             Id = Guid.NewGuid().ToString();
             DateTime = dateTime;
@@ -11,6 +11,6 @@
 
         public string Id { get; set; }
         public DateTime DateTime { get; set; }
-        public double Quotation { get; set; } 
+        public float Quotation { get; set; } 
     }
 }
